Automatic Meme Content Generator/requirements.txt
flask
praw
python-dotenv
requests
Flask-SQLAlchemy
pyttsx3
